Porting 18564 to msft.202405.
