+++ tests/snappi_tests/pfc/test_m2o_oversubscribe_lossless_lossy.py
@@ -14,6 +14,7 @@ from tests.snappi_tests.pfc.files.m2o_oversubscribe_lossless_lossy_helper import
 from tests.common.snappi_tests.snappi_test_params import SnappiTestParams            # noqa: F401
 from tests.snappi_tests.files.helper import setup_ports_and_dut, multidut_port_info  # noqa: F401
 from tests.common.snappi_tests.variables import pfcQueueGroupSize, pfcQueueValueDict        # noqa: F401
+from tests.snappi_tests.cisco.helper import disable_voq_watchdog                  # noqa: F401
 logger = logging.getLogger(__name__)
 pytestmark = [pytest.mark.topology('multidut-tgen', 'tgen')]
